backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Resume Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.0
openpyxl==3.1.2

# NLP and ML
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
nltk==3.8.1
gensim==4.3.2
transformers==4.35.2
torch==2.1.1

# Data processing
pandas==2.1.3
numpy==1.25.2

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1